pyproject.toml
[project]
name = "rt-disasterresponsedp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "kafka-python>=2.2.10",
    "pandas>=2.2.3",
    "psycopg2>=2.9.10",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "streamlit>=1.45.1",
]
